made the black brick background the background for the entire trivia game
