+++ app/(games)/trivia-game/[id]/game/page.tsx
@@ -1,7 +1,9 @@
-import React from 'react'
+import React from "react";
+
+// assets
+
+// components
 
 export default function page() {
-  return (
-    <div>Gameplay page for the trivia game</div>
-  )
+    return <div>Gameplay page for the trivia game</div>;
 }
