+++ app/(games)/trivia-game/[id]/lobby/page.tsx
@@ -16,18 +16,18 @@ import JoinStuff from "../../components/JoinStuff";
 // const bo1 = Black_Ops_One({ weight: "400", subsets: ["latin"] });
 
 export default function TriviaLobby() {
-    const bgStyling = {
-        backgroundImage: `url(${blackBrickBackground.src})`,
-        // <a href="https://www.freepik.com/free-vector/pink-blue-neon-frame-neon-frame-dark-background-vector_25519175.htm#query=neon%20design&position=29&from_view=keyword&track=ais&uuid=4dca27a6-d051-4271-8515-cbb20617a721">Image by rawpixel.com</a> on Freepik
-        // <a href="https://www.freepik.com/free-photo/black-brick-wall-textured-background_3475675.htm#query=black%20brick%20wall&position=1&from_view=search&track=ais&uuid=b7b18b50-2020-476b-95e9-4440c4d181eb">Image by rawpixel.com</a> on Freepik
-        backgroundSize: "100% 100%",
-        backgroundPosition: "center",
-        backgroundRepeat: "no-repeat",
-        padding: "7vh 5vw",
-    };
+    // const bgStyling = {
+    //     backgroundImage: `url(${blackBrickBackground.src})`,
+    //     // <a href="https://www.freepik.com/free-vector/pink-blue-neon-frame-neon-frame-dark-background-vector_25519175.htm#query=neon%20design&position=29&from_view=keyword&track=ais&uuid=4dca27a6-d051-4271-8515-cbb20617a721">Image by rawpixel.com</a> on Freepik
+    //     // <a href="https://www.freepik.com/free-photo/black-brick-wall-textured-background_3475675.htm#query=black%20brick%20wall&position=1&from_view=search&track=ais&uuid=b7b18b50-2020-476b-95e9-4440c4d181eb">Image by rawpixel.com</a> on Freepik
+    //     backgroundSize: "100% 100%",
+    //     backgroundPosition: "center",
+    //     backgroundRepeat: "no-repeat",
+    //     padding: "7vh 5vw",
+    // };
 
     return (
-        <div className={`h-screen bg-cover`} style={bgStyling}>
+        // <div className={`h-screen bg-cover`}>
             <div className={`flex flex-col justify-center h-full border-4 rounded-2xl ${styles.neonBorder}`}>
                 <div className="flex flex-row justify-evenly">
                     <div>
@@ -49,6 +49,6 @@ export default function TriviaLobby() {
                     </CoolButton>
                 </div>
             </div>
-        </div>
+        // </div>
     );
 }
